将Uri相关扩展方法移动至Uri.kt文件中
