+++ rxhttp-compiler/src/main/java/com/rxhttp/compiler/ClassHelper.kt
@@ -69,7 +69,7 @@ object ClassHelper {
             import rxhttp.wrapper.parse.Parser;
             import rxhttp.wrapper.parse.SimpleParser;
             import rxhttp.wrapper.parse.StreamParser;
-            import rxhttp.wrapper.utils.KotlinExtensions;
+            import rxhttp.wrapper.utils.UriUtil;
             import rxhttp.wrapper.utils.LogUtil;
 
             /**
@@ -221,7 +221,7 @@ object ClassHelper {
                                                                                                             
                 public final Observable<Uri> asAppendDownload(Context context, Uri uri, Scheduler scheduler,
                                                               Consumer<Progress> progressConsumer) {        
-                    long length = KotlinExtensions.length(uri, context);                                
+                    long length = UriUtil.length(uri, context);                                
                     if (length >= 0) setRangeHeader(length, -1, true);                                                                              
                     return asParser(StreamParser.get(context, uri), scheduler, progressConsumer);           
                 }                                                                                           
@@ -237,7 +237,7 @@ object ClassHelper {
                             Uri uri = uriFactory.query();
                             StreamParser<Uri> parser;
                             if (uri != null) {
-                                long length = KotlinExtensions.length(uri, uriFactory.getContext());
+                                long length = UriUtil.length(uri, uriFactory.getContext());
                                 if (length >= 0)
                                     setRangeHeader(length, -1, true);
                                 parser = StreamParser.get(uriFactory.getContext(), uri);
