+++ rxhttp/src/main/java/rxhttp/wrapper/callback/OutputStreamFactory.kt
@@ -1,9 +1,7 @@
 package rxhttp.wrapper.callback
 
-import android.content.ContentUris
 import android.content.Context
 import android.net.Uri
-import android.provider.MediaStore
 import okhttp3.Response
 import rxhttp.wrapper.OkHttpCompat
 import rxhttp.wrapper.entity.OutputStreamWrapper
@@ -75,27 +73,5 @@ private fun String.replaceSuffix(response: Response): String {
         format(OkHttpCompat.pathSegments(response).last())
     } else {
         this
-    }
-}
-
-//find the Uri by filename and relativePath, return null if find fail.  RequiresApi 29
-fun Uri.query(context: Context, filename: String?, relativePath: String?): Uri? {
-    if (filename.isNullOrEmpty() || relativePath.isNullOrEmpty()) return null
-    val realRelativePath = relativePath.let {
-        //Remove the prefix slash if it exists
-        if (it.startsWith("/")) it.substring(1) else it
-    }.let {
-        //Suffix adds a slash if it does not exist
-        if (it.endsWith("/")) it else "$it/"
-    }
-    val columnNames = arrayOf(
-        MediaStore.MediaColumns._ID,
-    )
-    return context.contentResolver.query(this, columnNames,
-        "relative_path=? AND _display_name=?", arrayOf(realRelativePath, filename), null)?.use {
-        if (it.moveToFirst()) {
-            val uriId = it.getLong(0)
-            ContentUris.withAppendedId(this, uriId)
-        } else null
     }
 }
