+++ rxhttp/src/main/java/rxhttp/wrapper/utils/Call.kt
@@ -1,0 +1,55 @@
+package rxhttp.wrapper.utils
+
+import kotlinx.coroutines.suspendCancellableCoroutine
+import okhttp3.Call
+import okhttp3.Callback
+import okhttp3.Response
+import rxhttp.wrapper.parse.Parser
+import java.io.IOException
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+
+/**
+ * User: ljx
+ * Date: 2020/9/13
+ * Time: 22:41
+ */
+internal suspend fun Call.await(): Response {
+    return suspendCancellableCoroutine { continuation ->
+        continuation.invokeOnCancellation {
+            cancel()
+        }
+        enqueue(object : Callback {
+
+            override fun onResponse(call: Call, response: Response) {
+                continuation.resume(response)
+            }
+
+            override fun onFailure(call: Call, e: IOException) {
+                continuation.resumeWithException(e)
+            }
+        })
+    }
+}
+
+internal suspend fun <T> Call.await(parser: Parser<T>): T {
+    return suspendCancellableCoroutine { continuation ->
+        continuation.invokeOnCancellation {
+            cancel()
+        }
+        enqueue(object : Callback {
+
+            override fun onResponse(call: Call, response: Response) {
+                try {
+                    continuation.resume(parser.onParse(response))
+                } catch (t: Throwable) {
+                    continuation.resumeWithException(t)
+                }
+            }
+
+            override fun onFailure(call: Call, e: IOException) {
+                continuation.resumeWithException(e)
+            }
+        })
+    }
+}
